# Step 44: Unify assembly loop and introduce an optimisation