For quite some time I've felt that step-44's way of assembly, namely having one method to assemble the system matrix and another the RHS, was not such a good approach. It seemed the correct thing to do at the time (the code that step-44 was based off of had a very expensive tangent matrix assembly), but in retrospect it makes extending the code to MPI and other scenarios difficult. In the last few years, each time that I've used this tutorial as a base code I've ended up unifying these two functions. So it seems like its worth formally introducing this approach upstream so that no-one has to perform this task again. It also then better aligns with the other tutorials in this respect.

I've also added a nice little optimisation that @davydden pointed out to me some years ago. Its a small change but it nearly halves the assembly time for the default setup, and it looks like its even more impactful as the number of elements increases.
| Global refinements | Old implementation (tangent + RHS) | New implementation (unified assm. + opt.) |
|--------------------|--------------------------------------|----------------------|
|                             2 |                                         2.65s + 0.5s |                           1.25s |
|                             3 |                                         15s + 0.95s |                            5.3s |

For other coupled problems involving rank-[`>=2`] tensors, the same trick gives even more reward. I can't believe that I had never noticed it before, even though I've stared at this code for many hours.

On the correctness of the change, here is the output at the last time step for the current implementation
```
Timestep 10 @ 1.000e+00s
___________________________________________________________________________________________________________________________________________________________
                 SOLVER STEP                   |  LIN_IT   LIN_RES    RES_NORM     RES_U     RES_P      RES_J     NU_NORM      NU_U       NU_P       NU_J
___________________________________________________________________________________________________________________________________________________________
  0  ASM_R  ASM_K  CST  ------  SLV  PP  UQPH  |     456  1.905e-06  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00
  1  ASM_R  ASM_K  CST  ------  SLV  PP  UQPH  |     260  2.967e-04  1.544e-01  1.544e-01  1.154e+13  1.849e+06  6.013e-02  7.397e-02  6.013e-02  6.013e-02
  2  ASM_R  ASM_K  CST  ------  SLV  PP  UQPH  |     279  2.099e-06  2.903e-03  2.903e-03  6.974e+10  2.060e+03  2.732e-03  1.436e-03  2.732e-03  2.733e-03
  3  ASM_R  ASM_K  CST  ------  SLV  PP  UQPH  |     280  2.494e-09  2.086e-06  2.086e-06  4.357e+07  4.103e+00  1.594e-06  6.952e-07  1.594e-06  1.595e-06
  4  ASM_R  CONVERGED!
___________________________________________________________________________________________________________________________________________________________
Relative errors:
Displacement:	6.952e-07
Force: 		9.451e-10
Dilatation:	1.528e-06
v / V_0:	1.000e-09 / 1.000e-09 = 1.000e+00
```
and this is after the change
```
Timestep 10 @ 1.000e+00s
______________________________________________________________________________________________________________________________________________________
               SOLVER STEP                |  LIN_IT   LIN_RES    RES_NORM     RES_U     RES_P      RES_J     NU_NORM      NU_U       NU_P       NU_J
______________________________________________________________________________________________________________________________________________________
  0  CST  ASM_SYS  ------  SLV  PP  UQPH  |     456  1.905e-06  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00  1.000e+00
  1  CST  ASM_SYS  ------  SLV  PP  UQPH  |     260  2.967e-04  1.544e-01  1.544e-01  1.154e+13  1.849e+06  6.013e-02  7.397e-02  6.013e-02  6.013e-02
  2  CST  ASM_SYS  ------  SLV  PP  UQPH  |     279  2.099e-06  2.903e-03  2.903e-03  6.974e+10  2.060e+03  2.732e-03  1.436e-03  2.732e-03  2.733e-03
  3  CST  ASM_SYS  ------  SLV  PP  UQPH  |     280  2.494e-09  2.086e-06  2.086e-06  4.357e+07  4.103e+00  1.594e-06  6.952e-07  1.594e-06  1.595e-06
  4  CST  ASM_SYS  CONVERGED!
______________________________________________________________________________________________________________________________________________________
Relative errors:
Displacement:	6.952e-07
Force: 		9.451e-10
Dilatation:	1.528e-06
v / V_0:	1.000e-09 / 1.000e-09 = 1.000e+00
```
The results look the same to me, and the amended table output also looks OK.