@jppelteret 2020-10-26
> Do you now assemble the matrix when it is not needed or do you always need both anyways? In other situations I have a single assemble(bool only_rhs) function to be more efficient while still having the code unified.

Thanks for your comment and suggestion! The only time that the tangent matrix is not needed is if we attain equilibrium, and then exit the NR loop without solving the linear system. So that's one unnecessary assembly of the matrix per timestep. But, of course, one can only know that once the residual is constructed, and so one could end up assembling the RHS without the tangent matrix, finding out that the equilibrium solution has not been attained, and then constructing it again along with the tangent matrix. 

I agree that what you suggest is not hard to do (adding a flag is an elegant way to do this), but perhaps in this case it's not worth doing. My initial feeling is that it would also be another twist to an already complex tutorial. Also, it would be contradictory to the new documentation that I added which states come cases for which you need to assemble both together to get the correct residual. So then I'd need to explain that...