@jppelteret 2020-10-26
Many thanks for your comments @tjhei and @kronbichler! I've attended to them, and have also fixed a bug that I just discovered. I'll squash all commits at the end of the review.